--- core-ui/src/components/Clusters/components/KubeconfigUpload/KubeconfigTextArea/KubeconfigTextArea.scss
@@ -1,16 +1,0 @@
-.kubeconfig-text-area {
-  textarea {
-    min-height: 150px;
-    display: block;
-    margin-bottom: 16px;
-    resize: none;
-    width: 100%;
-    background: var(--sapButton_Background);
-    border-color: var(--sapNeutralBorderColor);
-    color: var(--sapTextColor);
-  }
-
-  button {
-    width: 100%;
-  }
-}
